refactor(module schemas): correct file names and type names
