+++ src/modules/Login/schemas/login.ts
@@ -5,4 +5,4 @@ export const loginSchema = z.object({
   password: z.string().min(5, 'login-form.errors.password-input.length'),
 });
 
-export type LoginFormInputsType = z.infer<typeof loginSchema>;
+export type loginSchema = z.infer<typeof loginSchema>;
